dom/tests/mochitest/bugs/test_bug823173.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=823173
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 823173</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=823173">Mozilla Bug 823173</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 823173 **/
try {
  ok(!(navigator instanceof Window), "navigator is not an instance of Window");
} catch (e) {
  ok(false, "instanceof tests should not throw");
}
</script>
</pre>
</body>
</html>
